+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 
 - support for chiplot format (preview feature only)
+- checkboxes in a preview to control y-axis scaling (linear or log) of the respective plots
 - persistence of state in a preview, such as checkbox and dropdown menu selections
   - Restore the state when a preview becomes visible after being moved into the background.
     - By default the state inside graphs such as scaling is not stored. When one wants to keep this state, enable `spec-data.preview.retainContextWhenHidden` option (but keep it in mind that this increases memory usage).